Move JWT token verification to plug module
